Skip temp-media files delete in CI

+++ utils/graphene/tests.py
@@ -47,7 +47,9 @@ class CommonSetupClassMixin:
         super().tearDownClass()
         # clear the temporary media files
         try:
-            shutil.rmtree(os.path.join(settings.BASE_DIR, TEST_MEDIA_ROOT), ignore_errors=True)
+            # NOTE: CI will clean itself
+            if os.environ.get('CI', '').lower() != 'true':
+                shutil.rmtree(os.path.join(settings.BASE_DIR, TEST_MEDIA_ROOT), ignore_errors=True)
         except FileNotFoundError:
             pass
